Outsource verboseoutput from gpusatutils.h
